% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cheat_sheet.R
\docType{data}
\name{example_cheat_sheet}
\alias{example_cheat_sheet}
\title{Example  cheat sheet}
\format{dataframe}
\usage{
example_cheat_sheet
}
\description{
Example Pre-defined markers for cell type need to provided to the algorithm.
}
\keyword{datasets}
