% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example  data}
\format{dataframe}
\usage{
example_data
}
\description{
Example expression data matrix. The data is preffered to be in natural scale with markers as columns and cells as rows.Note: Gene names should be set as row names and duplicates are not allowed. Missing values are not allowed within the expression matrix. Check example- head(example_data)
}
\keyword{datasets}
