% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_annotation.R
\name{imaap_annotation}
\alias{imaap_annotation}
\title{Cell type annotation}
\usage{
imaap_annotation(cell_prob, cheat_sheet)
}
\arguments{
\item{cell_prob}{Probabilty of the expression of cell types calculated function- \code{\link{imaap_prob}}}

\item{cheat_sheet}{A CSV file listing the the markers that are used to define a cell type. It needs to follow the defined template.}
}
\value{
Returns a cell type annotation for each cell at each level
}
\description{
Based on the probability calculated for each celltype, a cell-type annotation is assigned to each cell.
}
