% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_marker_drop.R
\name{imaap_marker_drop}
\alias{imaap_marker_drop}
\title{Annotation of cells that did not express one or more markers}
\usage{
imaap_marker_drop(cell_prob, cell_annotation, cheat_sheet)
}
\arguments{
\item{cell_prob}{Probabilty of expression of a cell type calculated using the function- \code{\link{imaap_prob}}}

\item{cell_annotation}{Output of \code{\link{imaap_annotation}}}

\item{cheat_sheet}{A CSV file listing the the markers that are used to define a cell type. It needs to follow the defined template.}
}
\value{
Returns a cell type annotation for cells that have lost one or more markers
}
\description{
When multiple markers are used for annotation of a cell type, it is possible that one or more markers are not expressed, this function identifies those cells and marks them.
}
