% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_model.R
\name{imaap_model}
\alias{imaap_model}
\title{Fitting a gaussian mixture model to the users data}
\usage{
imaap_model(data, SD = 3)
}
\arguments{
\item{data}{Dataframe of data (natural scale) containing samples as rows and markers as columns.}

\item{SD}{Standard Deviation. Default is set to 3. This determines to what extend of the data distribution is considered as positive to marker expression. If your data distribution is clearly not bi-modal, it is suggested to be conservative in the selection of an SD value.}
}
\value{
Probability scores for each marker and a deconvolved data matrix for calculating percent positivity of markers of interest.
}
\description{
Gaussian mixture model is fit to every marker in the dataset and a probaility score is given to the areas of uncertanity.
}
