% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_model_plot.R
\name{imaap_model_plot}
\alias{imaap_model_plot}
\title{Distribution plot with the gaussian model fit}
\usage{
imaap_model_plot(data, peaks_low, peaks_high, SD)
}
\arguments{
\item{data}{Dataframe of data (natural scale) containing samples as rows and markers as columns.}

\item{peaks_low}{Mean expression value of the left most distribution in the gaussian fit}

\item{peaks_high}{Mean expression value of the left most distribution in the gaussian fit}

\item{SD}{Standard Deviation. Default is set to 3. This determines to what extend of the data distribution is considered as positive to marker expression. If your data distribution is clearly not bi-modal, it is suggested to be conservative in the selection of an SD value.}
}
\value{
Plots
}
\description{
The fuction plots the distribution for every marker along with the gaussian model fit.
}
