% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_one_call_per_cell.R
\name{imaap_one_call_per_cell}
\alias{imaap_one_call_per_cell}
\title{Reduce to one cell type call per cell}
\usage{
imaap_one_call_per_cell(cell_calling)
}
\arguments{
\item{cell_calling}{The cell annotation output from \code{\link{imaap_marker_drop}}}
}
\value{
A single cell type calling for each cell in the dataset.
}
\description{
Every cell gets only one annotation; that is all levels of cell annotation are reduced to one.
}
