% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaap_prob.R
\name{imaap_prob}
\alias{imaap_prob}
\title{Probablility calculation for the user defined cell types.}
\usage{
imaap_prob(marker_prob, cheat_sheet)
}
\arguments{
\item{marker_prob}{Probabilty of the expression of individual markers calculared using the gaussian mixture modelling function- \code{\link{imaap_model}}}

\item{cheat_sheet}{A CSV file listing the the markers that are used to define a cell type. It needs to follow the defined template.}
}
\value{
Probability scores for each cell type of interest.
}
\description{
Takes in the probability of expression for individual marker expression and calculates the probability for cell types based on the user defined parameters.
}
